Add files via upload
You will visualize the mapped environment through the generated image
Instruction:

Check out this link for more information on the matplotlib C++ library. For information regarding the plot color and shape refer to the LineSpec and LineColor section of the MATLAB documentation.
Compiling:

$ ./compile.sh

Running:

$ ./main

Now, wait for the program to generate the map! then view it from 'Images' folder